Add files via upload
This file shows the influence of the severe Ischemia on NBCe1 activity. Sodium-potassium pump is inhibited about 80% to mimic the reduced supply of blood and oxygen. We observe the three distinct phase of the NBCe1 ( slow rise, reversal and depolarization). The first phase is due to the extracellular bicarbonate change, second phase is due to the potassium and the third phase is due to the potassium and Na+ influx. The reversibility disappears in the physiological range of the potassium,  supporting the hypothesis that NBCe1 reversal is dependent on the potassium change. Over all  we can see the strong Na+ load, infra-cellular acidifies, membrane depolarization and bicarbonate reversibility.