Add files via upload
This file contains all the necessary equations and the value of the parameters used in the model to generate the results. Bath concentrations of the respective ions were used for the extracellular elevations. We use mathematical functions to inhibit the NKA approximately 50% and 80%  to mimic transient and severe ischemia. Transient Ischemia is observed by inhibiting pump for approximately 2 minute during 30 minute simulations time. Severe Ischemia is observed by inhibiting the pump approximately 20 minutes. In transient Ischemia cell recovers after the 2 minute Ischemia is over. But in Intense Ischemia model predictions shows that cell dies and do not recover.